+++ MercadoPagoSDK/MercadoPagoSDK/PXBankDealComponentView.swift
@@ -1,0 +1,16 @@
+
+//
+//  PXBankDealComponentView.swift
+//  MercadoPagoSDK
+//
+//  Created by AUGUSTO COLLERONE ALFONSO on 11/5/18.
+//  Copyright © 2018 MercadoPago. All rights reserved.
+//
+
+import UIKit
+
+class PXBankDealComponentView: PXComponentView {
+    var imageView: UIView?
+    var titleLabel: UILabel?
+    var subtitleLabel: UILabel?
+}
